FAI-432: added skeleton component for double bar chart, renamed new features scores chart
