+++ ui-packages/packages/trusty/src/components/Molecules/SkeletonDoubleBarChart/SkeletonDoubleBarChart.scss
@@ -1,0 +1,27 @@
+.skeleton-double-barchart {
+  opacity: 0.8;
+  margin-top: 2rem;
+  &__chart {
+    height: 100%;
+    &--left {
+      padding: 0 10% 0 20%;
+    }
+    &--right {
+      padding: 0 20% 0 10%;
+    }
+    &__legend {
+      height: 15%;
+      text-align: center;
+    }
+    &__bars {
+      position: relative;
+      height: 85%;
+      .skeleton__stripe {
+        position: absolute;
+        display: block;
+        height: 1.5em;
+        margin: 0;
+      }
+    }
+  }
+}
